+++ config/app.php
@@ -183,8 +183,8 @@ return [
         App\Providers\VaporUiServiceProvider::class,
         App\Providers\ViewServiceProvider::class,
         \App\Providers\ControlServiceProvider::class,
-        \App\Providers\AdditionalAttributeProvider::class,
-        \App\Providers\GoogleDriveServiceProvider::class
+        \App\Providers\GoogleDriveServiceProvider::class,
+        \App\Providers\SettingServiceProvider::class
     ],
 
     /*