+++ resources/js/utilities/api/api.js
@@ -22,6 +22,9 @@ import ModuleInstanceServices from './resources/moduleinstanceservices';
 import Setting from './resources/setting';
 import ActionInstanceHistory from './resources/actioninstancehistory';
 import Progress from'./resources/progress';
+import OwnedResource from './resources/owned-resource';
+import Activities from './resources/activities';
+import ActivityEvaluation from './resources/activity-evaluation';
 
 import axios from 'axios';
 
@@ -51,6 +54,19 @@ export default class {
             return Promise.reject(error);
         });
         return this;
+    }
+
+
+    ownedResource() {
+        return new OwnedResource(this._http);
+    }
+
+    activities() {
+        return new Activities(this._http);
+    }
+
+    activityEvaluation() {
+        return new ActivityEvaluation(this._http);
     }
 
     action() {