+++ resources/views/partials/footer.blade.php
@@ -1,5 +1,5 @@
 <footer class="text-muted py-5" style=" bottom: 0; width: 100%; height: 5%; margin-top: 10px; margin-bottom: 2px; ">
     <div class="container" style="text-align: center;">
-        {!! siteSetting('pageText.footer', 'Portal Footer') !!}
+        {!! \App\Settings\Appearance\Messaging\Footer::getValue() !!}
     </div>
 </footer>